remove round and item and list visual refactor

+++ app/src/main/java/com/etologic/mahjongscoring2/data_source/local_data_source/local/daos/RoundsDao.kt
@@ -16,7 +16,7 @@ interface RoundsDao {
     fun insertOne(round: Round): Single<Long>  //TODO: probar con tests. ¿Hace falta la exception o devolvería un 0?
     
     @Query("SELECT * FROM Rounds WHERE gameId = :gameId AND roundId = :roundId")
-    fun getOne(gameId: Long, roundId: Long): Single<Round>
+    fun getOne(gameId: Long, roundId: Int): Single<Round>
     
     @Query("SELECT * FROM Rounds WHERE gameId = :gameId")
     fun getAllByGame(gameId: Long): Single<List<Round>>
@@ -25,7 +25,7 @@ interface RoundsDao {
     fun updateOne(round: Round): Single<Int>
     
     @Query("DELETE FROM Rounds WHERE gameId = :gameId AND roundId = :roundId")
-    fun deleteOne(gameId: Long, roundId: Long): Single<Int>
+    fun deleteOne(gameId: Long, roundId: Int): Single<Int>
     
     @Query("DELETE FROM Rounds WHERE gameId = :gameId")
     fun deleteByGame(gameId: Long): Single<Int>
